refactor(circular-deps): break metamask-template-renderer ↔ metamask-translation cycle
The 3-way cycle ran:
  safe-component-list → metamask-translation → metamask-template-renderer → safe-component-list

Break it with pure file/import reorganization:
- Remove the static `MetaMaskTranslation` entry from `safe-component-list.js`.
  `safeComponentList` is now mutated as a side-effect at the bottom of
  `metamask-translation.js` once the component is defined.
- `metamask-translation.js` now imports the renderer's file directly
  (`metamask-template-renderer.js`) instead of going through the
  package index. This avoids reintroducing the cycle through `index.js`.
- The renderer's `index.js` adds a side-effect import of
  `../metamask-translation` so the registration is wired up for any
  consumer that goes through the package entry.
- Update `snap-ui-renderer.js` to import the renderer from the package
  index (instead of the file path directly), so it also triggers the
  registration side-effect.

No runtime change — `safeComponentList.MetaMaskTranslation` is populated
before the renderer ever looks anything up, just via a side-effect now
instead of a static literal.
