+++ .madgerc
@@ -24,7 +24,6 @@
   },
   "allowedCircularGlob": [
     "ui/ducks/**",
-    "ui/store/**",
-    "ui/components/app/**"
+    "ui/store/**"
   ]
 }
