+++ development/circular-deps.jsonc
@@ -7,12 +7,5 @@
 // - For more information contact the Extension Platform team.
 
 [
-  [
-    "ui/components/app/metamask-template-renderer/index.js",
-    "ui/components/app/metamask-template-renderer/metamask-template-renderer.js",
-    "ui/components/app/metamask-template-renderer/safe-component-list.js",
-    "ui/components/app/metamask-translation/index.js",
-    "ui/components/app/metamask-translation/metamask-translation.js"
-  ],
   ["ui/ducks/metamask/metamask.js", "ui/store/actions.ts"]
 ]
