+++ ui/components/app/metamask-template-renderer/index.js
@@ -1,2 +1,4 @@
+import '../metamask-translation';
+
 export { default } from './metamask-template-renderer';
 export { SectionShape } from './section-shape';
